Confirmacion de eliminacion

+++ src/app/leibnitz/PIII/modulos/grupo3/grupo3.component.ts
@@ -21,9 +21,10 @@ export class Grupo3Component implements OnInit {
     })
   }
   deletePelicula(id: number) {
-    this.servicio.deletePelicula(id).subscribe(() => {
-      this.getPeliculas();
-    })
+    if(confirm('¿Quieres eliminar esta pelicula?')){
+      this.servicio.deletePelicula(id).subscribe(() => {
+        this.getPeliculas();
+      })
+    }
   }
-
 }
